requirements.txt
requests_html
beautifulsoup4
PyYAML
